feat: add address type flags for correct Stellar address reconstruction
Bytes32 encoding cannot preserve Stellar address types (Account G... vs
Contract C...). This adds explicit boolean type flags to ensure addresses
are correctly reconstructed on the destination chain.

Changes:
- Add receiver_is_account to Intent and IntentData structs
- Add repayment_is_account to FillRecord struct
- Update payload format to include flags (192 bytes total)
- Add CreateIntentParams struct for Soroban (avoids 10-param limit)
- Change messenger_id from u8 to u32 in Soroban (SDK limitation)
- Fix Soroban sha256 API usage and XDR imports
- Update all tests for new struct formats

All tests pass:
- Soroban: 9/9 tests
- EVM: 31/31 tests

+++ stellar/contracts/rozo-intents/src/events.rs
@@ -65,7 +65,7 @@ pub fn emit_fill_and_notify_sent(
     intent_id: BytesN<32>,
     relayer: Address,
     repayment_address: BytesN<32>,
-    messenger_id: u8,
+    messenger_id: u32,
 ) {
     env.events().publish(
         (Symbol::new(env, "fill_and_notify_sent"), intent_id),
@@ -77,7 +77,7 @@ pub fn emit_retry_notify_sent(
     env: &Env,
     intent_id: BytesN<32>,
     relayer: Address,
-    messenger_id: u8,
+    messenger_id: u32,
 ) {
     env.events().publish(
         (Symbol::new(env, "retry_notify_sent"), intent_id),
@@ -131,7 +131,7 @@ pub fn emit_relayer_removed(env: &Env, relayer: Address) {
         .publish((Symbol::new(env, "relayer_removed"),), relayer);
 }
 
-pub fn emit_messenger_adapter_set(env: &Env, messenger_id: u8, adapter: Address) {
+pub fn emit_messenger_adapter_set(env: &Env, messenger_id: u32, adapter: Address) {
     env.events()
         .publish((Symbol::new(env, "messenger_adapter_set"),), (messenger_id, adapter));
 }
